reducer created for cart
